Let palette Default mean what the effect says it means
Round 2, R2-2 and R2-3, both of them WLED's odd palette rule and both proved on
the device.

Palette 0 is not a palette on WLED. loadPalette() replaces it with the palette
the running effect declared, or with Party when it declared none
(FX_fcn.cpp:234 and :618). Here it was always Party, so Fire 2012 on "Default"
rendered Party where the device renders its own palette 35. The segment carries
a default_palette now, set on every effect change exactly as setMode() sets it.

The other half is the same three lines read the other way. setMode() writes the
segment's palette only when the metadata names one, so selecting an effect that
declares none leaves the palette the user is on alone; 160 of the 223 effects
here declare none. The engine reset all of them to 0. Read back off the device
with fxdef: true, selecting Fire 2012 then Ripple then Colortwinkles leaves the
palette on 35 all the way through, and Glitter, which declares 11, changes it.
That is what this does now. The select entity follows the engine, so its
published state follows with it.

m12 and si get upstream's treatment as well: m12 is reset to M12_PIXELS when
the metadata does not name it and si is left alone, and both carry upstream's
constrain() into the parser.

Round 2 reported "104 of 223 declare pal=" and "112 of 216 reset the palette".
The behaviour is right and the counts are not: stock WLED declares pal= on 59
of its 216 strings and this build on 63 of 223. The larger number is a device
readback taken during a capture run, where the palette an effect inherits is
whatever the effect before it left.

Also in wf_registry.cpp, because it is the same table: the two standard slider
labels are "Effect speed" and "Effect intensity", which is what WLED's own UI
restores for a "!" label (data/index.js:1646). That is the text the "Effect
controls" sensor publishes. No entity is renamed.

+++ tools/sim/main.cpp
@@ -486,8 +486,10 @@ int main(int argc, char **argv) {
       char name[64];
       effect_name(*entry.second, name, sizeof(name));
       const EffectDefaults d = effect_defaults(*entry.second);
-      printf("%-10s %-20s flags=0x%02X pal=%u sx=%u ix=%u m12=%u\n", entry.first.c_str(), name, d.flags, d.palette,
-             d.speed, d.intensity, d.map1d2d);
+      // pal and m12 print -1 when the metadata declares neither, which is what
+      // upstream's extractModeDefaults() returns and what the engine acts on.
+      printf("%-10s %-20s flags=0x%02X pal=%d sx=%u ix=%u m12=%d\n", entry.first.c_str(), name, d.flags,
+             static_cast<int>(d.palette), d.speed, d.intensity, static_cast<int>(d.map1d2d));
     }
     printf("%zu effect(s)\n", selected.size());
     return 0;
